Converted RSC from init.d to systemd managed app
Converted the init.d scripts to a systemd service file. Thus
our RSC will now be restarted if it crash (after 2.5sec delay).

Added a rsc user and group for our project. The home folder is
/home/rsc. Added a couple of utility scripts to the home folder.

Added a file in profile.d to add the sbin folders to the PATH
variable for our rsc user.

Added a DEFAULT_TIMEZONE variable to the machine config to set
our timezone to Africa/Johannesburg.